+++ Assets/CommonCoreGame/RPGGame/RPG/CharacterModel.cs
@@ -9,6 +9,8 @@ using CommonCore.Config;
 using System.Runtime.Serialization;
 using CommonCore.World;
 using CommonCore.State;
+using System.Collections;
+using System.Linq;
 
 namespace CommonCore.RpgGame.Rpg
 {
@@ -111,7 +113,32 @@ namespace CommonCore.RpgGame.Rpg
 
         public InventoryModel Inventory { get; private set; }
         public Dictionary<EquipSlot, int> AmmoInMagazine { get; set; }
-        public Dictionary<EquipSlot, InventoryItemInstance> Equipped { get; private set; }
+
+        [JsonIgnore]
+        public IDictionary<EquipSlot, InventoryItemInstance> Equipped { get; private set; }
+        [JsonProperty(PropertyName = "Equipped")]
+        private Dictionary<EquipSlot, InventoryItemInstance> EquippedJsonParseable //hack for parsing old saves and old RPG defs
+        {
+            set
+            {
+                if (value != null && value.Count > 0)
+                {
+                    Debug.LogWarning("Loading equipped items through legacy object-ref parsing");
+                    foreach (var kvp in value)
+                    {
+                        if (kvp.Value.InstanceUID == 0)
+                        {
+                            Debug.LogError($"Cannot equip item instance ({kvp.Value}) because it has no UID");
+                            continue;
+                        }
+                        EquippedIDs[kvp.Key] = kvp.Value.InstanceUID;
+                    }
+                }
+
+            }
+        }
+        [JsonProperty]
+        private Dictionary<EquipSlot, long> EquippedIDs { get; set; } = new Dictionary<EquipSlot, long>();
 
         //mostly for addons/game-specific stuff
         [JsonProperty(ItemTypeNameHandling = TypeNameHandling.All)]
@@ -124,7 +151,7 @@ namespace CommonCore.RpgGame.Rpg
 
             Inventory = new InventoryModel();
             Conditions = new List<Condition>();
-            Equipped = new Dictionary<EquipSlot, InventoryItemInstance>();
+            Equipped = new EquippedDictionaryProxy(this);
             AmmoInMagazine = new Dictionary<EquipSlot, int>();
             ExtraData = new Dictionary<string, object>();
 
@@ -524,6 +551,92 @@ namespace CommonCore.RpgGame.Rpg
         bool IPlayerFlagsSource.Contains(string flag)
         {
             return false; //nop for now
+        }
+
+
+        //how equipped items are now handled
+        private class EquippedDictionaryProxy : IDictionary<EquipSlot, InventoryItemInstance>
+        {
+            private CharacterModel CharacterModel;
+
+            public EquippedDictionaryProxy(CharacterModel characterModel)
+            {
+                CharacterModel = characterModel;
+            }
+
+            public InventoryItemInstance this[EquipSlot key] { 
+                get => CharacterModel.Inventory.GetItem(CharacterModel.EquippedIDs[key]);
+                set => CharacterModel.EquippedIDs[key] = value.InstanceUID; 
+            }
+
+            public ICollection<EquipSlot> Keys => CharacterModel.EquippedIDs.Keys;
+
+            public ICollection<InventoryItemInstance> Values => CharacterModel.EquippedIDs.Values.Select(v => CharacterModel.Inventory.GetItem(v)).ToArray();
+
+            public int Count => CharacterModel.EquippedIDs.Count;
+
+            public bool IsReadOnly => false;
+
+            public void Add(EquipSlot key, InventoryItemInstance value)
+            {
+                CharacterModel.EquippedIDs.Add(key, value.InstanceUID);
+            }
+
+            public void Add(KeyValuePair<EquipSlot, InventoryItemInstance> item)
+            {
+                CharacterModel.EquippedIDs.Add(item.Key, item.Value.InstanceUID);
+            }
+
+            public void Clear()
+            {
+                CharacterModel.EquippedIDs.Clear();
+            }
+
+            public bool Contains(KeyValuePair<EquipSlot, InventoryItemInstance> item)
+            {
+                throw new NotImplementedException();
+            }
+
+            public bool ContainsKey(EquipSlot key)
+            {
+                return CharacterModel.EquippedIDs.ContainsKey(key);
+            }
+
+            public void CopyTo(KeyValuePair<EquipSlot, InventoryItemInstance>[] array, int arrayIndex)
+            {
+                throw new NotImplementedException();
+            }
+
+            public IEnumerator<KeyValuePair<EquipSlot, InventoryItemInstance>> GetEnumerator()
+            {
+                return CharacterModel.EquippedIDs.Select(kvp => new KeyValuePair<EquipSlot, InventoryItemInstance>(kvp.Key, CharacterModel.Inventory.GetItem(kvp.Value))).GetEnumerator();
+            }
+
+            public bool Remove(EquipSlot key)
+            {
+                return CharacterModel.EquippedIDs.Remove(key);
+            }
+
+            public bool Remove(KeyValuePair<EquipSlot, InventoryItemInstance> item)
+            {
+                throw new NotImplementedException();
+            }
+
+            public bool TryGetValue(EquipSlot key, out InventoryItemInstance value)
+            {
+                if(CharacterModel.EquippedIDs.TryGetValue(key, out long id))
+                {
+                    value = CharacterModel.Inventory.GetItem(id);
+                    return true;
+                }
+                value = default;
+                return false;
+            }
+
+            IEnumerator IEnumerable.GetEnumerator()
+            {
+                return this.GetEnumerator();
+            }
         }