+++ Assets/CommonCoreGame/RPGGame/World/Player/PlayerMovementComponent.cs
@@ -665,7 +665,7 @@ namespace CommonCore.RpgGame.World
                     Debug.LogWarning("Player out of bounds, teleporting!");
 
                     Vector3 teleportPoint = sceneController.WorldBounds.center + new Vector3(0, 2f, 0);
-                    GameObject spawnPoint = WorldUtils.FindObjectByTID("DefaultPlayerSpawn");
+                    GameObject spawnPoint = WorldUtils.FindPlayerSpawn();
                     if (spawnPoint != null)
                     {
                         teleportPoint = spawnPoint.transform.position;